# openapi.yaml
openapi: 3.0.0
info:
  title: Blog API Documentation
  version: 1.0.0
  description: | 
    Blog API is a RESTful backend service for modern blogging platforms.
    It provides secure authentication, role-based access, and CRUD operations for blog content.


    🚀 **Features**:
      - User authentication and authorization using JWT
      - Role-based access control (Admin / User)
      - CRUD operations for blog posts
      - Protected routes with middleware
      - Pagination and filtering support
      - Well-documented REST APIs with Swagger UI

    🛠 **Tech Stack**:
      - Node.js & Express.js
      - MongoDB with Mongoose ODM
      - RESTful API architecture
      - JWT-based authentication & authorization
      - Swagger UI (OpenAPI 3.0) for API documentation
      - Environment configuration using dotenv
      - Deployed on Render

  contact: 
    name: Sai Kowshik 
    email: saikowshik0306@gmail.com 
    url: https://github.com/03-Bunny-06/blog-api
servers:
  - url: https://blog-api-cc47.onrender.com
...
security: 
  - AdminAuth: []
  - UserAuth: []
paths:
...
  /user/signup:
    post:
      tags:
        - Users
      summary: SignUp
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example: ''
      parameters:
        - name: username
          in: header
          schema:
            type: string
          example: number4.dev
        - name: password
          in: header
          schema:
            type: string
          example: n4.dev123
      responses:
        '200':
          description: OK
          headers:
            Date:
              schema:
                type: string
                example: Thu, 29 Jan 2026 05:34:12 GMT
            Content-Type:
              schema:
                type: string
                example: application/json; charset=utf-8
            Content-Length:
              schema:
                type: integer
                example: '40'
            Connection:
              schema:
                type: string
                example: keep-alive
            CF-RAY:
              schema:
                type: string
                example: 9c565bed2c2979ef-HYD
            Content-Encoding:
              schema:
                type: string
                example: br
            etag:
              schema:
                type: string
                example: W/"24-GS9ZkRlQWaaM0laoy5lry1MO7TU"
            rndr-id:
              schema:
                type: string
                example: 20ee6f0a-f60a-47f0
            vary:
              schema:
                type: string
                example: Accept-Encoding
            x-powered-by:
              schema:
                type: string
                example: Express
            x-render-origin-server:
              schema:
                type: string
                example: Render
            cf-cache-status:
              schema:
                type: string
                example: DYNAMIC
            Server:
              schema:
                type: string
                example: cloudflare
            alt-svc:
              schema:
                type: string
                example: h3=":443"; ma=86400
          content:
            text/plain:
              schema:
                type: string
              example: |-
                {
                    "msg": "User created successfully!"
                }
        '400':
          description: Bad Request
          headers:
            Date:
              schema:
                type: string
                example: Thu, 29 Jan 2026 05:33:19 GMT
            Content-Type:
              schema:
                type: string
                example: application/json; charset=utf-8
            Content-Length:
              schema:
                type: integer
                example: '55'
            Connection:
              schema:
                type: string
                example: keep-alive
            CF-RAY:
              schema:
                type: string
                example: 9c565a1b7ba4d81d-HYD
            Content-Encoding:
              schema:
                type: string
                example: br
            etag:
              schema:
                type: string
                example: W/"33-f1WyNTRqh79sjnCXJ/SfXPag7GY"
            rndr-id:
              schema:
                type: string
                example: 6eb06906-352b-4a71
            vary:
              schema:
                type: string
                example: Accept-Encoding
            x-powered-by:
              schema:
                type: string
                example: Express
            x-render-origin-server:
              schema:
                type: string
                example: Render
            cf-cache-status:
              schema:
                type: string
                example: DYNAMIC
            Server:
              schema:
                type: string
                example: cloudflare
            alt-svc:
              schema:
                type: string
                example: h3=":443"; ma=86400
          content:
            text/plain:
              schema:
                type: string
              example: |-
                {
                    "msg": "User already exists try (SignIn) instead!"
                }
  /user/signin:
    post:
      tags:
        - Users
      summary: SignIn
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                username: User 2
                password: password2
      parameters:
        - name: username
          in: header
          schema:
            type: string
          example: number1.dev
        - name: password
          in: header
          schema:
            type: string
          example: n1.dev123
      responses:
        '200':
          description: OK
          headers:
            Date:
              schema:
                type: string
                example: Thu, 29 Jan 2026 05:35:13 GMT
            Content-Type:
              schema:
                type: string
                example: application/json; charset=utf-8
            Content-Length:
              schema:
                type: integer
                example: '165'
            Connection:
              schema:
                type: string
                example: keep-alive
            CF-RAY:
              schema:
                type: string
                example: 9c565d697c487b7a-HYD
            Content-Encoding:
              schema:
                type: string
                example: br
            etag:
              schema:
                type: string
                example: W/"b9-qkjlyi+GZ14NAOkjMiurzs77mCE"
            rndr-id:
              schema:
                type: string
                example: 0b1de6b0-f4ef-4e53
            vary:
              schema:
                type: string
                example: Accept-Encoding
            x-powered-by:
              schema:
                type: string
                example: Express
            x-render-origin-server:
              schema:
                type: string
                example: Render
            cf-cache-status:
              schema:
                type: string
                example: DYNAMIC
            Server:
              schema:
                type: string
                example: cloudflare
            alt-svc:
              schema:
                type: string
                example: h3=":443"; ma=86400
          content:
            text/plain:
              schema:
                type: string
              example: |-
                {
                    "msg": "Signed in successfully!!!",
                    "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6Im51bWJlcjEuZGV2IiwiaWF0IjoxNzY5NjY0OTEzfQ.so-Jl2xF8u7mYRy_MyNTwyPiqvlvE-1Dgnti9wAEEnU"
                }
        '400':
          description: Bad Request
          headers:
            Date:
              schema:
                type: string
                example: Thu, 29 Jan 2026 05:35:53 GMT
            Content-Type:
              schema:
                type: string
                example: application/json; charset=utf-8
            Content-Length:
              schema:
                type: integer
                example: '43'
            Connection:
              schema:
                type: string
                example: keep-alive
            CF-RAY:
              schema:
                type: string
                example: 9c565e6518817b7a-HYD
            Content-Encoding:
              schema:
                type: string
                example: br
            etag:
              schema:
                type: string
                example: W/"27-yobfxnp0hg5Gf/iV24945IQkR2U"
            rndr-id:
              schema:
                type: string
                example: ee3c305d-dfab-4432
            vary:
              schema:
                type: string
                example: Accept-Encoding
            x-powered-by:
              schema:
                type: string
                example: Express
            x-render-origin-server:
              schema:
                type: string
                example: Render
            cf-cache-status:
              schema:
                type: string
                example: DYNAMIC
            Server:
              schema:
                type: string
                example: cloudflare
            alt-svc:
              schema:
                type: string
                example: h3=":443"; ma=86400
          content:
            text/plain:
              schema:
                type: string
              example: |-
                {
                    "msg": "Invalid username or password!"
                }
...
  /user/search:
    get:
      tags:
        - Users
      summary: Blog Topic Based Filtering
      parameters:
        - name: t
          in: query
          schema:
            type: string
          example: science
      responses:
        '200':
          description: OK
          headers:
            Date:
              schema:
                type: string
                example: Thu, 29 Jan 2026 05:46:45 GMT
            Content-Type:
              schema:
                type: string
                example: application/json; charset=utf-8
            Content-Length:
              schema:
                type: integer
                example: '461'
            Connection:
              schema:
                type: string
                example: keep-alive
            CF-RAY:
              schema:
                type: string
                example: 9c566e50680321ba-HYD
            Content-Encoding:
              schema:
                type: string
                example: br
            etag:
              schema:
                type: string
                example: W/"99c-7OhxMAuHGadlRiQE7ciNGX2INRA"
            rndr-id:
              schema:
                type: string
                example: e21d4a44-2bde-4ae0
            vary:
              schema:
                type: string
                example: Accept-Encoding
            x-powered-by:
              schema:
                type: string
                example: Express
            x-render-origin-server:
              schema:
                type: string
                example: Render
            cf-cache-status:
              schema:
                type: string
                example: DYNAMIC
            Server:
              schema:
                type: string
                example: cloudflare
            alt-svc:
              schema:
                type: string
                example: h3=":443"; ma=86400
          content:
            text/plain:
              schema:
                type: string
              examples:
                example-0:
                  summary: Successful Response For Tech Topic Term
                  value: |-
                    {
                        "searchedTopicData": [
                            {
                                "_id": "6929865e3de2a4f78e6eed45",
                                "title": "How to become a Python Full Stack Dev in (2025)",
                                "description": "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s.",
                                "imageLink": "https://media.geeksforgeeks.org/wp-content/uploads/20240415204701/How-to-Become-a-Python-Full-Stack-Developer.png",
                                "topic": "tech",
                                "__v": 0
                            },
                            {
                                "_id": "6929866b3de2a4f78e6eed47",
                                "title": "How to become a Full Stack Dev in (2025)",
                                "description": "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s.",
                                "imageLink": "https://media.geeksforgeeks.org/wp-content/uploads/20240415204701/How-to-Become-a-Python-Full-Stack-Developer.png",
                                "topic": "tech",
                                "__v": 0
                            },
                            {
                                "_id": "692986bb823e75e591570385",
                                "title": "How to become a Java Full Stack Dev in (2025)",
                                "description": "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s.",
                                "imageLink": "https://media.geeksforgeeks.org/wp-content/uploads/20240415204701/How-to-Become-a-Python-Full-Stack-Developer.png",
                                "topic": "tech",
                                "__v": 0
                            },
                            {
                                "_id": "69298761823e75e591570389",
                                "title": "How to make coding more interesting",
                                "description": "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s.",
                                "imageLink": "https://media.geeksforgeeks.org/wp-content/uploads/20240415204701/How-to-Become-a-Python-Full-Stack-Developer.png",
                                "topic": "tech",
                                "__v": 0
                            },
                            {
                                "_id": "69763c765afda2dca12c2a3e",
                                "title": "Tech that changed the dynamics of 2025!",
                                "description": "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s.",
                                "imageLink": "https://media.geeksforgeeks.org/wp-content/uploads/20240415204701/How-to-Become-a-Python-Full-Stack-Developer.png",
                                "topic": "tech",
                                "__v": 0
                            },
                            {
                                "_id": "697a01541d612a03813bf1ce",
                                "title": "How to get a good tech job in 2026?",
                                "description": "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s.",
                                "imageLink": "https://media.geeksforgeeks.org/wp-content/uploads/20240415204701/How-to-Become-a-Python-Full-Stack-Developer.png",
                                "topic": "tech",
                                "__v": 0
                            }
                        ]
                    }
                example-1:
                  summary: Successful Response For Science Topic Term
                  value: |-
                    {
                        "searchedTopicData": [
                            {
                                "_id": "69298713823e75e591570387",
                                "title": "Will the sun explode in 5 billion years",
                                "description": "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s.",
                                "imageLink": "https://media.geeksforgeeks.org/wp-content/uploads/20240415204701/How-to-Become-a-Python-Full-Stack-Developer.png",
                                "topic": "science",
                                "__v": 0
                            },
                            {
                                "_id": "6929879c823e75e59157038b",
                                "title": "Is ATLAS 3I an interstellar comet or an UFO",
                                "description": "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s.",
                                "imageLink": "https://media.geeksforgeeks.org/wp-content/uploads/20240415204701/How-to-Become-a-Python-Full-Stack-Developer.png",
                                "topic": "science",
                                "__v": 0
                            },
                            {
                                "_id": "69298805823e75e59157038f",
                                "title": "Did NASA really land on moon or is it",
                                "description": "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s.",
                                "imageLink": "https://media.geeksforgeeks.org/wp-content/uploads/20240415204701/How-to-Become-a-Python-Full-Stack-Developer.png",
                                "topic": "science",
                                "__v": 0
                            },
                            {
                                "_id": "6929b06d8c75ac139670d7fb",
                                "title": "The Great Tech Infrastructure of 2025!",
                                "description": "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s.",
                                "imageLink": "https://media.geeksforgeeks.org/wp-content/uploads/20240415204701/How-to-Become-a-Python-Full-Stack-Developer.png",
                                "topic": "science",
                                "__v": 0
                            },
                            {
                                "_id": "697641935afda2dca12c2a42",
                                "title": "Does Aliens really exist ?",
                                "description": "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s.",
                                "imageLink": "https://media.geeksforgeeks.org/wp-content/uploads/20240415204701/How-to-Become-a-Python-Full-Stack-Developer.png",
                                "topic": "science",
                                "__v": 0
                            }
                        ]
                    }
        '404':
          description: Not Found
          headers:
            Date:
              schema:
                type: string
                example: Thu, 29 Jan 2026 05:45:04 GMT
            Content-Type:
              schema:
                type: string
                example: application/json; charset=utf-8
            Content-Length:
              schema:
                type: integer
                example: '60'
            Connection:
              schema:
                type: string
                example: keep-alive
            CF-RAY:
              schema:
                type: string
                example: 9c566bda8e0c4e87-HYD
            Content-Encoding:
              schema:
                type: string
                example: br
            etag:
              schema:
                type: string
                example: W/"3a-XJ3vnjGo5dBRvaO3lM6GHUwrWp4"
            rndr-id:
              schema:
                type: string
                example: 7cc553cf-73b7-40f7
            vary:
              schema:
                type: string
                example: Accept-Encoding
            x-powered-by:
              schema:
                type: string
                example: Express
            x-render-origin-server:
              schema:
                type: string
                example: Render
            cf-cache-status:
              schema:
                type: string
                example: DYNAMIC
            Server:
              schema:
                type: string
                example: cloudflare
            alt-svc:
              schema:
                type: string
                example: h3=":443"; ma=86400
          content:
            text/plain:
              schema:
                type: string
              example: |-
                {
                    "msg": "No search results found (or) Invalid topic term!"
                }
...
components:
  securitySchemes:
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        🔐 **Admin Authorization**

        - Use an **Admin JWT token.**
        - Paste the token in the **Authorize** button.
        - Format must be exactly: 
        `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6ImFkbWluMiIsImlhdCI6MTc2OTM1NjEwNH0.w2uHf-PySYP1V6wMZD48uPzZ_bf6a7wgE1YZa0aqWss`
    UserAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        🔐 **User Authorization**

        - Use an **User JWT token.**
        - Paste the token in the **Authorize** button.
        - Format must be exactly: 
        `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6Im51bWJlcjEuZGV2IiwiaWF0IjoxNzY5NjY0OTEzfQ.so-Jl2xF8u7mYRy_MyNTwyPiqvlvE-1Dgnti9wAEEnU`